+++ frontend/src/components/ScheduleAddModal/ScheduleAddModal.tsx
@@ -5,7 +5,7 @@ import { useState } from 'react';
 import { useQuery } from 'react-query';
 import { useRecoilValue } from 'recoil';
 
-import { usePostSchedule } from '@/hooks/@queries/schedules';
+import { usePostSchedule } from '@/hooks/@queries/schedule';
 import useControlledInput from '@/hooks/useControlledInput';
 import useValidateSchedule from '@/hooks/useValidateSchedule';
